gateway-routers: Change folder structure
